fix: token refresh and logout flow

+++ app/db/crud.py
@@ -17,7 +17,8 @@ def create_user(db: Session, username:str, email:str, password:str) -> models.Us
 def get_user_by_email(db: Session, email:str) -> models.User | None:
     return db.query(models.User).filter(models.User.email == email).first()
 
-def delete_refresh_token(db: Session, refresh_token:str):
+def delete_refresh_token(db: Session, refresh_token:str, userId:int):
     instance = db.query(models.RefreshToken).get(refresh_token)
-    if instance:
+    if instance and instance.user_id == userId:
         db.delete(instance)
+        db.commit()
